ingestion/requirements.txt
openai>=1.30.0
supabase>=2.4.0
pdfplumber>=0.11.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
